django_backend/requirements.txt
Django==5.0.1
Pillow==10.2.0
django-cors-headers==4.4.0
djangorestframework==3.14.0
python-decouple==3.8
django-crispy-forms==2.1
crispy-tailwind==0.5.0
